requirements.txt
requests
lxml
tabulate
progressbar2
click
setuptools
